(Refactor) Move blocktore options related stuff to blockstore_options.rs
