+++ ledger/src/lib.rs
@@ -15,6 +15,7 @@ pub mod blockstore_db;
 pub mod blockstore_meta;
 #[macro_use]
 pub mod blockstore_metrics;
+pub mod blockstore_options;
 pub mod blockstore_processor;
 pub mod builtins;
 pub mod genesis_utils;
